moving model back under lib
