+++ src/lib/etl/jobs/user_match_runner.ts
@@ -1,5 +1,5 @@
 import { runBackfills, runUpdates } from '../user_match_scrape'
-import { createEtlJob, updateEtlJob } from '../../../model/etl_jobs'
+import { createEtlJob, updateEtlJob } from '../../model/etl_jobs'
 import { argv } from 'yargs'
 import md5 from 'md5'
