+++ src/lib/model/events.ts
@@ -1,0 +1,6 @@
+import {insertIntoDatabase} from "../etl/utils";
+import {SITE_EVENTS} from "../constants";
+
+export async function createEvent(event) {
+    return insertIntoDatabase(event, SITE_EVENTS);
+}
